Add option to the 'build' command to build pages individually

+++ bireli/management/commands/build.py
@@ -3,7 +3,7 @@ import shutil
 
 from django.conf import settings
 from django.core.management import call_command
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandError
 from django.urls import reverse
 from django.test import Client
 
@@ -18,13 +18,23 @@ class Command(BaseCommand):
     help = 'Build static site output.'
     leave_locale_alone = True
 
+    def add_arguments(self, parser):
+        parser.add_argument('args', nargs='*')
+
     def handle(self, *args, **options):
         """
         Request pages and build output.
         """
-        if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
-            shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
-        os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
+        if args:
+            pages = args
+            invalid = [page for page in pages if page not in list(get_pages())]
+            if invalid:
+                raise CommandError(f"Invalid pages: {', '.join(invalid)}")
+        else:
+            pages = get_pages()
+            if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
+                shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
+            os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
         os.makedirs(settings.STATIC_ROOT, exist_ok=True)
         call_command(
             command_name='collectstatic',
@@ -33,13 +43,14 @@ class Command(BaseCommand):
             verbosity=0
         )
         client = Client()
-        for page in get_pages():
+        for page in pages:
             url = reverse('page', kwargs={'slug': page})
             response = client.get(url)
             if page == 'index':
                 output_dir = settings.SITE_OUTPUT_DIRECTORY
             else:
                 output_dir = os.path.join(settings.SITE_OUTPUT_DIRECTORY, page)
-                os.makedirs(output_dir)
+                if not os.path.exists(output_dir):
+                    os.makedirs(output_dir)
             with open(os.path.join(output_dir, 'index.html'), 'wb') as f:
                 f.write(response.content)
